+++ src/englishData/pronouns.js
@@ -1,4 +1,4 @@
-import { capitalizeFirst } from '../../helpers/utils';
+import { capitalizeFirst } from '../utils';
 import { _re, _ve } from './linkingVerbs';
 
 export const they = actor => actor.grammar.pronouns.subject;
@@ -23,3 +23,41 @@ export const Theirs = actor => capitalizeFirst(theirs(actor));
 
 export const themself = actor => actor.grammar.pronouns.reflexive;
 export const Themself = actor => capitalizeFirst(themself(actor));
+
+
+export const THEY_PRONOUNS = {
+  pronouns: {
+    subject: 'they',
+    object: 'them',
+    possessive: 'their',
+    possessiveStrong: 'theirs',
+    reflexive: 'themself',
+  },
+  isPlural: true,
+};
+
+export const HE_PRONOUNS = {
+  pronouns: {
+    subject: 'he',
+    object: 'him',
+    possessive: 'his',
+    possessiveStrong: 'his',
+    reflexive: 'himself',
+    usesPluralVerbs: false,
+  },
+  isPlural: false,
+};
+
+export const SHE_PRONOUNS = {
+  pronouns: {
+    subject: 'she',
+    object: 'her',
+    possessive: 'her',
+    possessiveStrong: 'hers',
+    reflexive: 'herself',
+    usesPluralVerbs: false,
+  },
+  isPlural: false,
+};
+
+export const ALL_PRONOUNS = [THEY_PRONOUNS, HE_PRONOUNS, SHE_PRONOUNS];
