+++ src/modules/Encounters/CityEncounter.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import Encounter from './Encounter';
-import { addTreat } from '../reducers/gameReducer';
+import { addTreat } from '../../reducers/gameReducer';
 
 
 class CityEncounter extends Encounter {
