+++ src/modules/Encounters/Encounter.js
