+++ src/modules/Encounters/MySpriteEncounter.js
@@ -2,14 +2,14 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import Encounter from './Encounter';
-import { getActiveSprite, interactWithSprite } from '../reducers/spriteReducer';
-import { addTreat } from '../reducers/gameReducer';
-import { getTrustLevel } from '../gameData/spriteTrust';
-import { randomChoice } from '../helpers/utils';
+import { getActiveSprite, interactWithSprite } from '../../reducers/spriteReducer';
+import { addTreat } from '../../reducers/gameReducer';
+import { getTrustLevel } from '../Sprite/spriteTrust';
+import { randomChoice } from '../../utils';
 
 import { PET_TEMPLATES, WATER_TEMPLATES, GROOM_TEMPLATES,
-  TREAT_TEMPLATES } from '../gameData/templates/spriteTemplates';
-import { SING_ACTIONS, SING_REACTIONS } from '../gameData/templates/singTemplates';
+  TREAT_TEMPLATES } from '../../englishData/templates/spriteTemplates';
+import { SING_ACTIONS, SING_REACTIONS } from '../../englishData/templates/singTemplates';
 
 class MySpriteEncounter extends Encounter {
   getInitialText() {
