+++ src/modules/Encounters/WildSpriteEncounter.js
@@ -2,14 +2,14 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import Encounter from './Encounter';
-import { addTreat } from '../reducers/gameReducer';
-import { befriendSprite } from '../reducers/spriteReducer';
-import { getTrustLevel, TRUST_LEVELS } from '../gameData/spriteTrust';
-import { randomChoice } from '../helpers/utils';
+import { addTreat } from '../../reducers/gameReducer';
+import { befriendSprite } from '../../reducers/spriteReducer';
+import { getTrustLevel, TRUST_LEVELS } from '../Sprite/spriteTrust';
+import { randomChoice } from '../../utils';
 
-import { generateSprite } from '../helpers/spriteHelpers';
+import { generateSprite } from '../Sprite/spriteHelpers';
 import { PET_TEMPLATES, WATER_TEMPLATES, TREAT_TEMPLATES, WAIT_TEMPLATES,
-  APPROACH_TEMPLATES } from '../gameData/templates/spriteTemplates';
+  APPROACH_TEMPLATES } from '../../englishData/templates/spriteTemplates';
 
 
 const addTrustAndDistance = (sprite, delta) => {
