+++ src/modules/Sprite/name/badList.js
