+++ src/modules/Sprite/spriteEncyclopedia.js
