+++ src/modules/Sprite/spriteTrust.js
