+++ src/utils.js
@@ -43,3 +43,15 @@ export function sRandomInt(seedArg, max, minArg) {
 export function capitalizeFirst(string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
 }
+
+// Pluralize a noun based on a count.
+export const _n = (singular, pluralArg) => {
+  const plural = pluralArg || `${singular}s`;
+  return count => (count === 1 ? singular : plural);
+};
+
+// Pluralize a verb based on a subject object.
+export const _v = (plural, singularArg) => {
+  const singular = singularArg || `${plural}s`;
+  return subject => (subject.grammar.isPlural ? plural : singular);
+};
